Move ChainReaction into Android Library module
