+++ app/build.gradle
@@ -32,8 +32,7 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation "com.android.support:design:${libs.supportLibVersion}"
     implementation "com.android.support:appcompat-v7:${libs.supportLibVersion}"
     implementation "com.android.support:recyclerview-v7:${libs.supportLibVersion}"
@@ -44,9 +43,12 @@ dependencies {
 
     // Kotlin AND Anko
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:${kotlin.kotlinVersion}"
-    implementation "org.jetbrains.anko:anko-sdk25:${kotlin.ankoVersion}"// sdk23 also available
-    implementation "org.jetbrains.anko:anko-support-v4:${kotlin.ankoVersion}" // support-v4 bindings
-    implementation "org.jetbrains.anko:anko-appcompat-v7:${kotlin.ankoVersion}" // appcompat-v7 bindings
+    // sdk23 also available
+    implementation "org.jetbrains.anko:anko-sdk25:${kotlin.ankoVersion}"
+    // support-v4 bindings
+    implementation "org.jetbrains.anko:anko-support-v4:${kotlin.ankoVersion}"
+    // appcompat-v7 bindings
+    implementation "org.jetbrains.anko:anko-appcompat-v7:${kotlin.ankoVersion}"
 
     // FIREBASE AND MAPS
     implementation "com.google.firebase:firebase-core:${libs.firebaseLibVersion}"
@@ -76,6 +78,7 @@ dependencies {
     // RxJava
     implementation 'io.reactivex.rxjava2:rxjava:2.1.1'
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
+    implementation project(':chainreaction')
 }
 
 // ADD THIS AT THE BOTTOM
