--- app/src/main/java/au/com/beba/chainReaction/demo/ReactorImplDemo.kt
@@ -1,86 +1,0 @@
-package au.com.beba.chainReaction.demo
-
-import au.com.beba.chainReaction.feature.logger.ConsoleLogger
-import au.com.beba.chainReaction.feature.chain.ChainCallback
-import au.com.beba.chainReaction.feature.chain.ChainTask
-import au.com.beba.chainReaction.feature.reactor.withPhases.BaseChainWithPhases
-
-class LoggingInChain : BaseChainWithPhases() {
-    override val TAG = LoggingInChain::class.java.simpleName
-
-    override fun getChainTask(): ChainTask {
-        return object : ChainTask {
-            override fun run(callback: ChainTask.ChainTaskCallback) {
-                val status = ChainCallback.Status.SUCCESS
-                val taskResult = "L"
-                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
-                Thread.sleep(800)
-                callback.onResult(this, status, taskResult)
-            }
-        }
-    }
-}
-
-class SignInChain : BaseChainWithPhases() {
-    override val TAG = SignInChain::class.java.simpleName
-
-    override fun getChainTask(): ChainTask {
-        return object : ChainTask {
-            override fun run(callback: ChainTask.ChainTaskCallback) {
-                val status = ChainCallback.Status.SUCCESS
-                val taskResult = "S"
-                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
-                Thread.sleep(1000)
-                callback.onResult(this, status, taskResult)
-            }
-        }
-    }
-}
-
-class GetAccountsChain : BaseChainWithPhases() {
-    override val TAG = GetAccountsChain::class.java.simpleName
-
-    override fun getChainTask(): ChainTask {
-        return object : ChainTask {
-            override fun run(callback: ChainTask.ChainTaskCallback) {
-                val status = ChainCallback.Status.SUCCESS
-                val taskResult = "S-GA"
-                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
-                Thread.sleep(300)
-                callback.onResult(this, status, taskResult)
-            }
-        }
-    }
-}
-
-class GetCardsChain : BaseChainWithPhases() {
-    override val TAG = GetCardsChain::class.java.simpleName
-
-    override fun getChainTask(): ChainTask {
-        return object : ChainTask {
-            override fun run(callback: ChainTask.ChainTaskCallback) {
-                val status = ChainCallback.Status.SUCCESS
-                val taskResult = "S-GC"
-                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
-                Thread.sleep(400)
-                callback.onResult(this, status, taskResult)
-            }
-        }
-    }
-}
-
-class PostSignInChain : BaseChainWithPhases() {
-    override val TAG = PostSignInChain::class.java.simpleName
-
-    override fun getChainTask(): ChainTask {
-        return object : ChainTask {
-            override fun run(callback: ChainTask.ChainTaskCallback) {
-                val status = ChainCallback.Status.SUCCESS
-                val taskResult = "PS"
-                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
-                Thread.sleep(1500)
-                callback.onResult(this, status, taskResult)
-            }
-        }
-    }
-}
