--- app/src/main/java/au/com/beba/chainReaction/feature/chain/Chain.kt
@@ -1,16 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface Chain {
-    fun addToChain(chain: Chain)
-
-    fun startChain(callback: ChainCallback)
-
-    fun preMainTaskPhase()
-    fun mainTaskPhase()
-    fun postMainTaskPhase()
-    fun getChainTask(): ChainTask
-
-    fun getChainResult(): Any?
-    fun getChainStatus(): ChainCallback.Status
-    fun setChainStatus(newStatus: ChainCallback.Status)
-}
