--- app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainDecision.kt
@@ -1,5 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface ChainDecision {
-    fun decision(links: List<Chain>, chain : ChainDecisionListener)
-}
