--- app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainDecisionListener.kt
@@ -1,6 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface ChainDecisionListener {
-    fun onDecisionDone(finalStatus: ChainCallback.Status)
-    fun onDecisionNotDone()
-}
