--- app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainTask.kt
@@ -1,9 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface ChainTask {
-    interface ChainTaskCallback {
-        fun onResult(task: ChainTask, newStatus: ChainCallback.Status, taskResult: Any?)
-    }
-
-    fun run(callback: ChainTaskCallback)
-}
