--- app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainWithPhases.kt
@@ -1,11 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface ChainWithPhases : Chain {
-    fun addReaction(reaction: Reaction)
-
-    // PHASES
-    fun linksPhase()
-
-    fun reactionsPhase()
-    fun decisionPhase()
-}
