--- app/src/main/java/au/com/beba/chainReaction/feature/chain/Reactor.kt
@@ -1,5 +1,0 @@
-package au.com.beba.chainReaction.feature.chain
-
-interface Reactor {
-    val chainDecision: ChainDecision
-}
