+++ chainreaction/build.gradle
@@ -1,0 +1,56 @@
+def libs = rootProject.ext.libs
+def kotlin = rootProject.ext.kotlin
+def test = rootProject.ext.testing
+
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+
+android {
+    compileSdkVersion 27
+
+
+
+    defaultConfig {
+        minSdkVersion 21
+        targetSdkVersion 27
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.android.support:appcompat-v7:27.0.2'
+
+    // Kotlin AND Anko
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:${kotlin.kotlinVersion}"
+    // sdk23 also available
+    implementation "org.jetbrains.anko:anko-sdk25:${kotlin.ankoVersion}"
+    // support-v4 bindings
+    implementation "org.jetbrains.anko:anko-support-v4:${kotlin.ankoVersion}"
+    // appcompat-v7 bindings
+    implementation "org.jetbrains.anko:anko-appcompat-v7:${kotlin.ankoVersion}"
+
+}
+
+// TEST
+dependencies {
+    testImplementation "junit:junit:${test.junit}"
+    androidTestImplementation 'com.android.support.test:runner:1.0.1'
+    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+}
+
+repositories {
+    mavenCentral()
+}
