+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/ChainCallback.kt
@@ -1,0 +1,12 @@
+package au.com.beba.chainreaction.chain
+
+interface ChainCallback {
+    fun onDone(status: Status)
+
+    enum class Status {
+        NOT_STARTED,
+        IN_PROGRESS,
+        ERROR,
+        SUCCESS
+    }
+}
