+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/ChainDecision.kt
@@ -1,0 +1,5 @@
+package au.com.beba.chainreaction.chain
+
+interface ChainDecision {
+    fun decision(links: List<Chain>, chain: ChainDecisionListener)
+}
