+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/ChainDecisionListener.kt
@@ -1,0 +1,8 @@
+package au.com.beba.chainreaction.chain
+
+import au.com.beba.chainreaction.chain.ChainCallback
+
+interface ChainDecisionListener {
+    fun onDecisionDone(finalStatus: ChainCallback.Status)
+    fun onDecisionNotDone()
+}
