+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/ChainWithPhases.kt
@@ -1,0 +1,11 @@
+package au.com.beba.chainreaction.chain
+
+interface ChainWithPhases : Chain {
+    fun addReaction(reaction: Reaction)
+
+    // PHASES
+    fun linksPhase()
+
+    fun reactionsPhase()
+    fun decisionPhase()
+}
