+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/Reaction.kt
@@ -1,0 +1,3 @@
+package au.com.beba.chainreaction.chain
+
+class Reaction(val type: String, val task: (Any?) -> (Any?))
