+++ chainreaction/src/main/java/au/com/beba/chainreaction/chain/Reactor.kt
@@ -1,0 +1,5 @@
+package au.com.beba.chainreaction.chain
+
+interface Reactor {
+    val chainDecision: ChainDecision
+}
