+++ chainreaction/src/test/java/au/com/beba/chainreaction/ReactorImplDemo.kt
@@ -1,0 +1,91 @@
+package au.com.beba.chainreaction
+
+import au.com.beba.chainreaction.chain.ChainCallback
+import au.com.beba.chainreaction.chain.ChainTask
+import au.com.beba.chainreaction.logger.ConsoleLogger
+import au.com.beba.chainreaction.reactor.BaseChainWithPhases
+
+class LoggingInChain : BaseChainWithPhases() {
+    @Suppress("PropertyName")
+    override val TAG: String = LoggingInChain::class.java.simpleName
+
+    override fun getChainTask(): ChainTask {
+        return object : ChainTask {
+            override fun run(callback: ChainTask.ChainTaskCallback) {
+                val status = ChainCallback.Status.SUCCESS
+                val taskResult = "L"
+                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
+                Thread.sleep(800)
+                callback.onResult(this, status, taskResult)
+            }
+        }
+    }
+}
+
+class SignInChain : BaseChainWithPhases() {
+    @Suppress("PropertyName")
+    override val TAG: String = SignInChain::class.java.simpleName
+
+    override fun getChainTask(): ChainTask {
+        return object : ChainTask {
+            override fun run(callback: ChainTask.ChainTaskCallback) {
+                val status = ChainCallback.Status.SUCCESS
+                val taskResult = "S"
+                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
+                Thread.sleep(1000)
+                callback.onResult(this, status, taskResult)
+            }
+        }
+    }
+}
+
+class GetAccountsChain : BaseChainWithPhases() {
+    @Suppress("PropertyName")
+    override val TAG: String = GetAccountsChain::class.java.simpleName
+
+    override fun getChainTask(): ChainTask {
+        return object : ChainTask {
+            override fun run(callback: ChainTask.ChainTaskCallback) {
+                val status = ChainCallback.Status.SUCCESS
+                val taskResult = "S-GA"
+                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
+                Thread.sleep(300)
+                callback.onResult(this, status, taskResult)
+            }
+        }
+    }
+}
+
+class GetCardsChain : BaseChainWithPhases() {
+    @Suppress("PropertyName")
+    override val TAG: String = GetCardsChain::class.java.simpleName
+
+    override fun getChainTask(): ChainTask {
+        return object : ChainTask {
+            override fun run(callback: ChainTask.ChainTaskCallback) {
+                val status = ChainCallback.Status.SUCCESS
+                val taskResult = "S-GC"
+                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
+                Thread.sleep(400)
+                callback.onResult(this, status, taskResult)
+            }
+        }
+    }
+}
+
+class PostSignInChain : BaseChainWithPhases() {
+    @Suppress("PropertyName")
+    override val TAG: String = PostSignInChain::class.java.simpleName
+
+    override fun getChainTask(): ChainTask {
+        return object : ChainTask {
+            override fun run(callback: ChainTask.ChainTaskCallback) {
+                val status = ChainCallback.Status.SUCCESS
+                val taskResult = "PS"
+                ConsoleLogger.log(TAG, "Running task with status=[%s] and result=[%s]".format(status, taskResult))
+                Thread.sleep(1500)
+                callback.onResult(this, status, taskResult)
+            }
+        }
+    }
+}
